requirements.txt
numpy>=1.21.0
pygame>=2.0.0
pyyaml>=6.0.0
PyQt5>=5.15.0
pandas>=1.3.0  # For data export capabilities